Add end of Legacy Support dates (new eol-legacy column)
The optional Legacy support add-on on top of Ubuntu Pro extends the
security maintenance and support for an additional 2 years, resulting in
12 years coverage overall. That information is already publicly present
in the Release Cycle page [1].

Add a new `eol-legacy` column for the legacy support and add the dates
for the supported LTS version. The Release Cycle page [1] mentions only
the month. So just add two years to the ESM end of life dates and pick
the next day that falls in the Tuesday to Thursday range.

[1] https://ubuntu.com/about/release-cycle
LP: #2073707